refactor: move the create wallet page header into its own component
